+++ lib/sparsify.rb
@@ -3,6 +3,7 @@
 require 'sparsify/version'
 require 'sparsify/utility_methods'
 require 'sparsify/guard_methods'
+require 'sparsify/separator'
 require 'sparsify/core_ext/kernel'
 
 # Provides sparse-key access to a Hash.
@@ -39,7 +40,7 @@ module Sparsify
   # Used internally by both Sparsify::Utility#sparse and
   # Sparsify::Utility#unsparse
   #
-  # @overload sparse_eachrm (options = {}, &block)
+  # @overload sparse_each(options = {}, &block)
   #   Yields once per key in sparse version of itself.
   #   @param options (see #sparse)
   #   @yieldparam [(sparse_key,value)]
@@ -74,8 +75,8 @@ module Sparsify
   # @overload sparse_get(sparse_key, options = {})
   #   @param options (see #sparse)
   #   @return [Object] at that address or nil if none found
-  def sparse_fetch(*args, &block)
-    Sparsify.sparse_fetch(self, *args, &block)
+  def sparse_get(*args)
+    Sparsify.sparse_get(self, *args)
   end
 
   # Returns a deeply-nested hash version of self.