+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.1.0 (2015-06-04)
+
+ - Added Sparsify::expand(), which expands a subset of the given sparse hash (@yaauie)
+ - Fixed extraneous characters in documentation (@yaauie)
+ - Actually provided Sparsify#sparse_get to line up with documentation (@yaauie)
+
 ## v1.0.1 (2013-12-22)
 
  - Made Sparsify() and Unsparsify() helpers private methods on Kernel instead